backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
numpy
pandas
tqdm
python-dotenv
openai
spacy
reportlab
python-multipart
pdfplumber
requests
pytesseract
Pillow
opencv-python

torch
scikit-learn